+++ channels/server/channels.c
@@ -51,8 +51,11 @@
 #include <freerdp/server/telemetry.h>
 #include <freerdp/server/rdpgfx.h>
 #include <freerdp/server/disp.h>
-#include <freerdp/server/camera-device-enumerator.h>
-#include <freerdp/server/camera-device.h>
+
+#if defined(CHANNEL_RDPECAM_SERVER)
+#include <freerdp/server/rdpecam-enumerator.h>
+#include <freerdp/server/rdpecam.h>
+#endif
 
 #ifdef WITH_CHANNEL_GFXREDIR
 #include <freerdp/server/gfxredir.h>
@@ -79,8 +82,10 @@ void freerdp_channels_dummy(void)
 	TelemetryServerContext* telemetry;
 	RdpgfxServerContext* rdpgfx;
 	DispServerContext* disp;
+#if defined (CHANNEL_RDPECAM_SERVER)
 	CamDevEnumServerContext* camera_enumerator;
 	CameraDeviceServerContext* camera_device;
+#endif
 #ifdef WITH_CHANNEL_GFXREDIR
 	GfxRedirServerContext* gfxredir;
 #endif // WITH_CHANNEL_GFXREDIR
@@ -110,10 +115,14 @@ void freerdp_channels_dummy(void)
 	rdpgfx_server_context_free(rdpgfx);
 	disp = disp_server_context_new(NULL);
 	disp_server_context_free(disp);
+
+#if defined (CHANNEL_RDPECAM_SERVER)
 	camera_enumerator = cam_dev_enum_server_context_new(NULL);
 	cam_dev_enum_server_context_free(camera_enumerator);
 	camera_device = camera_device_server_context_new(NULL);
 	camera_device_server_context_free(camera_device);
+#endif
+
 #ifdef WITH_CHANNEL_GFXREDIR
 	gfxredir = gfxredir_server_context_new(NULL);
 	gfxredir_server_context_free(gfxredir);
