+++ include/config/config.h.in
@@ -62,9 +62,6 @@
 #cmakedefine CHANNEL_AUDIN
 #cmakedefine CHANNEL_AUDIN_CLIENT
 #cmakedefine CHANNEL_AUDIN_SERVER
-#cmakedefine CHANNEL_CAMERA
-#cmakedefine CHANNEL_CAMERA_CLIENT
-#cmakedefine CHANNEL_CAMERA_SERVER
 #cmakedefine CHANNEL_CLIPRDR
 #cmakedefine CHANNEL_CLIPRDR_CLIENT
 #cmakedefine CHANNEL_CLIPRDR_SERVER
@@ -95,9 +92,6 @@
 #cmakedefine CHANNEL_PRINTER
 #cmakedefine CHANNEL_PRINTER_CLIENT
 #cmakedefine CHANNEL_PRINTER_SERVER
-#cmakedefine CHANNEL_RDPEI
-#cmakedefine CHANNEL_RDPEI_CLIENT
-#cmakedefine CHANNEL_RDPEI_SERVER
 #cmakedefine CHANNEL_RAIL
 #cmakedefine CHANNEL_RAIL_CLIENT
 #cmakedefine CHANNEL_RAIL_SERVER
@@ -107,6 +101,9 @@
 #cmakedefine CHANNEL_RDPDR
 #cmakedefine CHANNEL_RDPDR_CLIENT
 #cmakedefine CHANNEL_RDPDR_SERVER
+#cmakedefine CHANNEL_RDPECAM
+#cmakedefine CHANNEL_RDPECAM_CLIENT
+#cmakedefine CHANNEL_RDPECAM_SERVER
 #cmakedefine CHANNEL_RDPEI
 #cmakedefine CHANNEL_RDPEI_CLIENT
 #cmakedefine CHANNEL_RDPEI_SERVER
